chore: Improve Bot section of README
In order to run bots the server-ip.txt file must be set, otherwise Zwift will throw error logs similar to "ERROR] Error connecting to TCP host 172.18.0.2:3025 [60] Operation timed out". This should be set to localhost in the default set-up provided in the readme since it's either running on localhost, or all the ports are being exposed on the docker-container.

This commit extends the "Bots" section of the README in order to explain this.